fix: #92 still includes siblings in folder
Storing `dirname` erases too much information and we cant distinguish siblings. So given:

```
/folder1/
/folder1/foo.ts
/folder1/bar.ts
[...]
/folder2/
/folder2/bat.ts
[...]
```

foo.zip gets both foo and bar, and so does bar.zip, and so on.

This change assumes that there may be generated files that are not listed in `buildResults`, but that they are "prefix-named", that is `foo.js` => `foo.js.map` - generalizing, we ignore here `foo.js*`