Always include files that are in patterns
Before this change if a directory or file was in node_modules for
example `.prisma` directory that gets created by prisma but not marked
as a dep or peer dep it would be removed.

Adding this directory to patterns did not include it in the artifact.
This change seeks to always include files that are present in patterns.