fix: don't set the chokidar option awaitWriteFinish
This option causes chokidar to wait for 2 seconds after a file has
changed before it emits the changed event. This causes build times to be
very slow which is impractical. The option is default off in chokidar
and their docs say:

    The appropriate duration setting is heavily dependent on the OS and
    hardware. For accurate detection this parameter should be relatively
    high, making file watching much less responsive. Use with caution.

Therefore, I don't think it's a good idea to set this option by default.
With the previous commit, chokidar options can be overridden, so a user
can set this option if necessary.