feat: adds support for `serverless-azure-functions` plugin
Adds support for Azure Function Apps built in Node via the
`serverless-azure-functions` plugin.

Currently when trying to use the plugin with the `azure`
runtime it does not compile because there is not a function
matcher available, only for AWS and GCP.

This commit adds a matcher for Azure, along with unit tests
and supporting documentation in README.md.