fix: spawning too many esbuilds builds (#491)
* Fix spawning multiple esbuilds on same file, that would cause memory usage to spike.
According to documentation of the changes in esbuild 0.17, calling rebuild on the same build params is “safe” in the sense that it would merge it into an existing build rather then spawn a new one.

Furthermore, the contexes of these rebuilds are now needed to be disposed of. That was the original issue that made me use build over the rebuild - in the initial import of esbuild 19, running serverless invoke and serverless package would never finish. Using build seemingly solved this, but it ended up causing the issue spawning too many builds described above.

I added a step on finishing invoking, deploying and packaging to make sure to dispose of the contexes.

* Use build if there is no context

* Proper naming