fix: preserve mtime when zipping with the node zipper
This means it's possible for code to retrieve the correct mtime, for
example to return in a `Last-Modified` header.

Since we have a test which verifies a checksum of a zip file created by
this function, we can be sure the results are deterministic. To support
this, update the test inputs to have specified mtimes.